requirements.txt
Django>=5.2
psycopg2-binary
python-decouple
djangorestframework
gunicorn
djangorestframework-simplejwt
reportlab
pillow
